+++ ios-runtime/src/iosMain/kotlin/net/multigesture/kanama/types/NodePath.kt
@@ -1,9 +1,0 @@
-package net.multigesture.kanama.types
-
-data class NodePath(val path: String) {
-  override fun toString(): String = path
-
-  companion object {
-    val EMPTY = NodePath("")
-  }
-}